third_party/blink/web_tests/platform/generic/svg/animations/animate-path-nested-transforms-expected.txt

SVG 1.1 dynamic animation tests

test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rootSVGElement.getBBox().x is 131
PASS rootSVGElement.getBBox().y is -90
PASS rootSVGElement.getBBox().x is 331
PASS rootSVGElement.getBBox().y is 550
PASS successfullyParsed is true

TEST COMPLETE
